Update file paths & move all string configuration to config.js
