+++ lib/config.js
@@ -1,21 +1,20 @@
+'use strict';
 /* --- SpoilerBot Strings --- */
-
-var botConfig = require("./config.json");
 
 class Config {
   constructor() {
     // Single Characters
-    this.PREFIX = botConfig.prefix;
-    this.DELINIATOR = botConfig.deliniator;
+    this.PREFIX = "!";
+    this.DELINIATOR = "|";
 
     // Styling
-    this.COLOR = botConfig.color;
+    this.COLOR = "#5b187c";
 
     // Commands
-    this.CMD_SPOILER = this.PREFIX + botConfig.spoiler_cmd;
+    this.CMD_SPOILER = this.PREFIX + "spoiler";
 
     // Bot Info
-      this.HOMEPAGE = "https://bobbysig.github.io/SpoilerBot/";
+    this.HOMEPAGE = "https://bobbysig.github.io/SpoilerBot/";
 
     // Help Messages
     this.HELP_TITLE = "SpoilerBot Help";
