--- lib/config.json
@@ -1,6 +1,0 @@
-{
-  "prefix": "!",
-  "deliniator": "|",
-  "spoiler_cmd": "spoiler",
-  "color": "#5b187c"
-}
