+++ lib/spoilerbot.js
@@ -1,6 +1,8 @@
+'use strict';
+
 const Discord = require("discord.js");
 const ROT13 = require('caesar-salad').ROT13;
-const Config = require("./config.js");
+const Config = require("../lib/config");
 
 class SpoilerBot {
   constructor(client, disableDefaultMsgBindings = false) {
