+++ package.json
@@ -5,8 +5,9 @@
   "main": "server.js",
   "scripts": {
     "test": "npm run bot-test-unit",
-    "bot-test-unit": "cross-env NODE_ENV=testing mocha 'bot/*.spec.js'",
-    "start": "node --use_strict server.js"
+    "bot-test-unit": "cross-env NODE_ENV=testing mocha './test/*.spec.js'",
+    "start": "node --use_strict server.js",
+    "main": "./lib/spoilerbot.js"
   },
   "dependencies": {
     "caesar-salad": "^1.0.1",
@@ -28,8 +29,7 @@
   },
   "license": "GPL-3.0-or-later",
   "keywords": [
-    "node",
-    "glitch",
-    "express"
+    "discord",
+    "bot"
   ]
 }
