+++ test/test-setup.spec.js
@@ -1,6 +1,5 @@
 // test-setup.spec.js
 const sinon = require('sinon');
-const chai = require('chai');
 
 afterEach(function () {
   sinon.restore();
